Complete User CRUD

+++ src/main/java/com/ailab/Planning/Poker/services/UserServices.java
@@ -1,33 +1,61 @@
 package com.ailab.Planning.Poker.services;
 
+import com.ailab.Planning.Poker.dto.UserDTO;
 import com.ailab.Planning.Poker.entity.User;
+import com.ailab.Planning.Poker.mapper.UserMapper;
 import com.ailab.Planning.Poker.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class UserServices {
     @Autowired
     UserRepository userRepository;
 
-    public List<User> getAll() {
-        return userRepository.findAll();
+    @Autowired
+    UserMapper userMapper;
+
+    public List<UserDTO> getAllUser() {
+        return userRepository.findAll()
+                .stream()
+                .map(userMapper::entityToDto)
+                .collect(Collectors.toList());
     }
 
-    public Optional<User> getByUsername(String username) {
-        return userRepository.findById(username);
+    public UserDTO getByUserName(String userName) {
+        return userMapper.entityToDto(userRepository.findById(userName).orElse(new User()));
     }
 
-    public User save(User newUser) {
-        return userRepository.save(newUser);
+    public UserDTO save(UserDTO newUser) {
+        User user = new User();
+        userMapper.dtoToEntity(newUser, user);
+        return userMapper.entityToDto(userRepository.save(user));
     }
 
-    public void delete(String username) {
-        userRepository.deleteById(username);
+    public UserDTO update(UserDTO newUser, String oldUsername) {
+        Optional<User> optionalUser = userRepository.findById(oldUsername);
+
+        if (optionalUser.isPresent()) {
+            User existingUser = optionalUser.get();
+            userMapper.dtoToEntity(newUser, existingUser);
+            userRepository.save(existingUser);
+            return userMapper.entityToDto(existingUser);
+        } else {
+            return null;
+        }
+    }
+
+    public HttpStatus delete(String username) {
+        if (userRepository.existsById(username)) {
+            userRepository.deleteById(username);
+            return HttpStatus.OK;
+        } else {
+            return HttpStatus.NOT_FOUND;
+        }
     }
 }
